Use new project structure also in test setup.
